refactor(utils): :truck: rename `rocv_io` to `io_helpers`
